Add error output and debug option
This commit adds a mechanism for both error output and a debug mode. If
a file named "error" exists within the cache directory, the contents are
output and the cache directory is not deleted.

If the "--debug" option is used, then the cache directory is left
intact whether or not an error occurred.