Removes unused function for escaping quotes
Function for escaping quotes was not used and so it has been removed.
Also, since functions defined in "fire" file are not usable in
subshells, this is an improper place to define function. Shared
functions should be defined in the "lib/shared" directory in the same
way that other functions are.